Move connect config to conf module
