+++ tests/dmap/test_functional.py
@@ -6,8 +6,8 @@ import ipaddress
 from tests.log_output_handler import LogOutputHandler
 from aiohttp.test_utils import (AioHTTPTestCase, unittest_run_loop)
 
-from pyatv import (AppleTV, AirPlayService, DmapService, connect_to_apple_tv,
-                   const, exceptions)
+from pyatv import (connect_to_apple_tv, const, exceptions)
+from pyatv.conf import (AirPlayService, DmapService, AppleTV)
 from pyatv.dmap import pairing
 from tests.fake_daap_atv import (
     FakeDaapAppleTV, AppleTVUseCases, DEVICE_PIN, DEVICE_CREDENTIALS)
